backend/chat/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

psycopg2-binary>=2.9.9
redis>=5.0.1
django-redis>=5.4.0

PyJWT>=2.8.0
requests>=2.31.0
aiohttp>=3.9.1

python-dotenv>=1.0.0

drf-spectacular>=0.27.0

pytest>=7.4.0
pytest-django>=4.7.0
pytest-asyncio>=0.23.0

pydantic>=2.0.0
msgpack>=1.0.5